requirements.txt

appdirs==1.4.3
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.15.0
alembic==1.7.6
charset-normalizer==2.0.12
cloudinary==1.29.0
confusable-homoglyphs==3.2.0
cryptography==37.0.1
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.7
django-allauth==0.50.0
django-bootstrap4==22.1
django-environ==0.8.1
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
django-registration==3.2
django-rest-auth==0.9.5
django-utils-six==2.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.3
oauthlib==3.2.0
phonenumbers==8.12.47
Pillow==9.1.0
psycopg2==2.9.3
psycopg2-binary==2.8.6
pycparser==2.21
PyJWT==2.3.0
python-decouple==3.6
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
urllib3==1.26.9
whitenoise==6.0.0